+++ src/main/java/org/withtime/be/withtimebe/WithTimeBeApplication.java
@@ -4,13 +4,21 @@ import java.util.TimeZone;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
+import org.withtime.be.withtimebe.domain.log.repository.VisitLogRepository;
 
 import jakarta.annotation.PostConstruct;
 
 @EnableScheduling
 @EnableJpaAuditing
+@EnableJpaRepositories(basePackages = {"org.withtime.be"}, excludeFilters = {
+    @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = VisitLogRepository.class)})
+@EnableMongoRepositories(basePackageClasses = VisitLogRepository.class)
 @SpringBootApplication
 public class WithTimeBeApplication {
