+++ src/main/java/org/withtime/be/withtimebe/domain/log/service/query/VisitLogQueryService.java
@@ -3,7 +3,7 @@ package org.withtime.be.withtimebe.domain.log.service.query;
 import java.time.LocalDate;
 import java.util.List;
 
-import org.withtime.be.withtimebe.domain.log.entity.VisitLog;
+import org.withtime.be.withtimebe.domain.log.model.VisitLog;
 
 public interface VisitLogQueryService {
 	List<VisitLog> findDailyVisitLogList();
