+++ src/main/java/org/withtime/be/withtimebe/domain/log/service/query/VisitLogQueryServiceImpl.java
@@ -2,12 +2,10 @@ package org.withtime.be.withtimebe.domain.log.service.query;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.withtime.be.withtimebe.domain.log.entity.VisitLog;
+import org.withtime.be.withtimebe.domain.log.model.VisitLog;
 import org.withtime.be.withtimebe.domain.log.repository.VisitLogRepository;
 
 import lombok.RequiredArgsConstructor;
